clean up bot slop
