+++ test/unit/lib/otel/metrics/nr-capturing-exporter.test.js
@@ -51,7 +51,10 @@ test.afterEach(async (ctx) => {
  * @returns {Promise<object>} The collected `ResourceMetrics`.
  */
 async function collect(ctx) {
-  ctx.nr.meterProvider.getMeter('test-meter').createCounter('test-counter').add(1, { foo: 'bar' })
+  ctx.nr.meterProvider
+    .getMeter('test-meter')
+    .createCounter('test-counter')
+    .add(1, { foo: 'bar' })
   const { resourceMetrics } = await ctx.nr.reader.collect()
   return resourceMetrics
 }
@@ -62,12 +65,11 @@ test('export serializes the metrics and reports success', async (t) => {
 
   let result = null
   exporter.export(metrics, (r) => { result = r })
-
-  // The exporter is synchronous: the callback fires before `export` returns.
   assert.deepEqual(result, { code: ExportResultCode.SUCCESS })
 
-  // The cached payload must match what the OTLP protobuf serializer produces.
-  const expected = Buffer.from(ProtobufMetricsSerializer.serializeRequest(metrics)).toString('base64')
+  const expected = Buffer.from(
+    ProtobufMetricsSerializer.serializeRequest(metrics)
+  ).toString('base64')
   assert.equal(exporter.lastSerialization, expected)
 })
 
@@ -77,7 +79,9 @@ test('export writes an audit log of the serialized payload', async (t) => {
 
   exporter.export(metrics, () => {})
 
-  const expected = Buffer.from(ProtobufMetricsSerializer.serializeRequest(metrics)).toString('base64')
+  const expected = Buffer.from(
+    ProtobufMetricsSerializer.serializeRequest(metrics)
+  ).toString('base64')
   assert.equal(t.nr.logs.length, 1)
   assert.deepEqual(t.nr.logs[0], [
     {
@@ -96,7 +100,6 @@ test('lastSerialization purges the cache on read', async (t) => {
   exporter.export(metrics, () => {})
 
   assert.notEqual(exporter.lastSerialization, '')
-  // Reading purges the cache so a subsequent harvest cannot re-flush stale data.
   assert.equal(exporter.lastSerialization, '')
 })
